+++ extensions/preprocess/page.php
@@ -32,18 +32,46 @@ function ultima_preprocess_page(&$vars) {
   // Tab logic
   // Allows custom rules for showing the page tabs.
   $vars['show_tabs'] = TRUE;
+  // Support for CCK fields which will allow toggling from node add/edit page
+  if (isset($vars['node']) && isset($vars['node']->field_show_tabs[0]['value'])) {
+    $vars['show_title'] = TRUE;
+    if ($vars['node']->field_show_title[LANGUAGE_NONE][0]['value'] == 0) {
+      $vars['show_title'] = FALSE;
+    }
+  }
 
   // Action links logic
   // Allows custom rules for showing the page action links.
   $vars['show_action_links'] = TRUE;
+  // Support for CCK fields which will allow toggling from node add/edit page
+  if (isset($vars['node']) && isset($vars['node']->field_show_action_links[0]['value'])) {
+    $vars['show_title'] = TRUE;
+    if ($vars['node']->field_show_title[LANGUAGE_NONE][0]['value'] == 0) {
+      $vars['show_title'] = FALSE;
+    }
+  }
 
   // Page title logic
   // Allows custom rules for showing the page title.
   $vars['show_title'] = TRUE;
+  // Support for CCK fields which will allow toggling from node add/edit page
+  if (isset($vars['node']) && isset($vars['node']->field_show_title[0]['value'])) {
+    $vars['show_title'] = TRUE;
+    if ($vars['node']->field_show_title[LANGUAGE_NONE][0]['value'] == 0) {
+      $vars['show_title'] = FALSE;
+    }
+  }
 
   // Page breadcrumbs logic
-  // Allows custom rules for showing the page title.
+  // Allows custom rules for showing the breadcrumbs.
   $vars['show_breadcrumbs'] = TRUE;
+  // Support for CCK fields which will allow toggling from node add/edit page
+  if (isset($vars['node']) && isset($vars['node']->field_show_breadcrumbs[0]['value'])) {
+    $vars['show_title'] = TRUE;
+    if ($vars['node']->field_show_title[LANGUAGE_NONE][0]['value'] == 0) {
+      $vars['show_title'] = FALSE;
+    }
+  }
 
   // Per content type processing
   if (isset($vars['node'])) {